requirements.txt
astroid==2.2.5
backcall==0.1.0
decorator==4.4.0
Django==2.2.3
django-filter==2.2.0
djangorestframework==3.10.1
ipython==7.6.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.14.0
lazy-object-proxy==1.4.1
mccabe==0.6.1
model-mommy==1.6.0
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
pylint==2.3.1
pytz==2019.1
rope==0.14.0
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
typed-ast==1.4.0
wcwidth==0.1.7
wrapt==1.11.2
